normal validation for verifySignature
In the corrected code, I've assigned the returned value of verifySignature to the isValid variable. If isValid is false, meaning the signature is invalid, an error is thrown and the request is rejected. This way, unauthorized users will not be able to send requests to your server with an incorrect secret key.

I assume that's it for now...